put interactions in their own file
