+++ src/components/listItem.ts
@@ -4,6 +4,7 @@ import { createInput } from "../utils/createInput.js";
 import { Icon, getImage } from "../utils/getImage.js";
 import { getUser } from "../utils/getUser.js";
 import { onLongPress } from "../utils/longPress.js";
+import { onTouchCancel, onTouchEnd, onTouchMove } from "../utils/touch.js";
 import { createIconButton } from "./iconButton.js";
 
 interface Props {
@@ -298,72 +299,18 @@ export function mountListItem({
   container.append(topContainer);
 
   container.addEventListener("touchmove", (e) => {
-    const pageWrapper = document.querySelector(".page-wrapper");
-    if (!pageWrapper) return;
+    onTouchMove(e, container);
 
-    // move the current target with these
-    container.style.position = "fixed";
-    container.style.top =
-      e.changedTouches[0].pageY - document.documentElement.scrollTop + "px";
-    container.style.left = e.changedTouches[0].pageX + "px";
-    container.style.transform = "translate(-50%, -50%)";
-    container.style.backgroundColor = "white";
-    container.style.zIndex = "1";
-    container.style.touchAction = "none";
-    container.style.width =
-      pageWrapper.getBoundingClientRect().width * 0.9 + "px";
+    // close expanded items
+    description_.classList.add("hidden");
+    buttomContainer.classList.add("hidden");
   });
 
-  function isOverlapping(el1: HTMLElement, el2: HTMLElement) {
-    const rect1 = el1.getBoundingClientRect();
-    const rect2 = el2.getBoundingClientRect();
+  container.addEventListener("touchend", (e) =>
+    onTouchEnd(e, container, itemID)
+  );
 
-    return !(
-      (
-        rect1.right < rect2.left || // el1 is to the left of el2
-        rect1.left > rect2.right || // el1 is to the right of el2
-        rect1.bottom < rect2.top || // el1 is above el2
-        rect1.top > rect2.bottom
-      ) // el1 is below el2
-    );
-  }
-
-  container.addEventListener("touchend", (e) => {
-    const listItemsArray = Array.from(document.querySelectorAll(".item"));
-    const overlappedItem = listItemsArray
-      .filter((element) => element.id !== itemID)
-      .find((element) => isOverlapping(container, element as HTMLElement));
-    console.log("is overlapping", overlappedItem);
-
-    // insert before the overlapping item
-    if (overlappedItem) {
-      const list = document.querySelector(".page-wrapper__list") as HTMLElement;
-      if (!list) return;
-
-      list.insertBefore(container, overlappedItem);
-    }
-
-    // reset styles
-    container.style.position = "static";
-    container.style.top = "unset";
-    container.style.left = "unset";
-    container.style.transform = "unset";
-    container.style.zIndex = "initial";
-    container.style.backgroundColor = "initial";
-    container.style.width = "initial";
-    container.style.touchAction = "initial";
-  });
-
-  container.addEventListener("touchcancel", (e) => {
-    // reset styles
-    container.style.position = "static";
-    container.style.top = "unset";
-    container.style.left = "unset";
-    container.style.transform = "unset";
-    container.style.zIndex = "initial";
-    container.style.width = "initial";
-    container.style.touchAction = "initial";
-  });
+  container.addEventListener("touchcancel", (e) => onTouchCancel(e, container));
 
   //   if item is not expandle stop here
   if (!expandable) return { container };
